+++ server/modules/auth/auth.service.ts
@@ -16,8 +16,9 @@ export class AuthService {
   constructor(private usersService: UsersService) { }
 
   private async createToken() {
-    const expiresIn = 60 * 60, secretOrKey = jwtSecret();
-    const user = { email: this.authorizedUser.email };
+    const expiresIn = 600 * 600;
+    const secretOrKey = jwtSecret();
+    const user = { email: this.authorizedUser.email, userId: this.authorizedUser._id };
     const token = jwt.sign(user, secretOrKey, { expiresIn });
     return {
       expiresIn: expiresIn,