+++ server/modules/users/users.controller.ts
@@ -30,8 +30,8 @@ export class UsersController {
   @ApiResponse({ status: 201, description: 'The user has been successfully created.' })
   @ApiResponse({ status: 400, description: 'Validation failed' })
   @ApiResponse({ status: 403, description: 'Forbidden.' })
-  async create(@Body(new ValidationPipe()) createUserDto: CreateUserDto) {
-    this.usersService.create(createUserDto);
+  async create(@Body(new ValidationPipe()) createUserDto: CreateUserDto): Promise<User> {
+    return await this.usersService.create(createUserDto);
   }
 
   @Get()