+++ server/modules/users/users.service.ts
@@ -13,13 +13,9 @@ export class UsersService {
 
   async create(userDto: CreateUserDto): Promise<User> {
     const createdUser = new this.userModel(userDto);
-    return await createdUser.save((err, user) => {
-      if (err) {
-        console.log(err);
-        throw new HttpException('Database error', HttpStatus.INTERNAL_SERVER_ERROR);
-      }
-      return user;
-    });
+    const user = await createdUser.save();
+    user.password = undefined;
+    return user;
   }
 
   async findAll(): Promise<User[]> {