+++ src/app/settings/settings.module.ts
@@ -16,8 +16,8 @@ import { InventoryService } from './inventory/inventory.service';
 import { CreateUserModalComponent } from './users/create-user-modal/create-user-modal.component';
 import { EditUserModalComponent } from './users/edit-user-modal/edit-user-modal.component';
 import { DeleteUserModalComponent } from './users/delete-user-modal/delete-user-modal.component';
-import { ValidateEqualDirective } from './users/validate-equal.directive';
 import { RolesService } from './roles/roles.service';
+import { IntegrationsComponent } from './integrations/integrations.component';
 
 @NgModule({
   imports: [
@@ -31,12 +31,12 @@ import { RolesService } from './roles/roles.service';
     SettingsComponent,
     UsersComponent,
     RolesComponent,
+    IntegrationsComponent,
     MainSettingsComponent,
     InventoryComponent,
     CreateUserModalComponent,
     EditUserModalComponent,
-    DeleteUserModalComponent,
-    ValidateEqualDirective
+    DeleteUserModalComponent
   ],
   providers: [
     UsersService,