+++ src/app/shared/shared.module.ts
@@ -2,6 +2,8 @@ import { NgModule } from '@angular/core';
 import { CommonModule } from '@angular/common';
 
 import { ClarityModule } from '@clr/angular';
+import { ValidateEqualDirective } from '../settings/users/validate-equal.directive';
+import { AutofocusDirective } from './autofocus.directive';
 
 @NgModule({
   imports: [
@@ -9,8 +11,13 @@ import { ClarityModule } from '@clr/angular';
     ClarityModule,
   ],
   exports: [
-    ClarityModule
+    ClarityModule,
+    ValidateEqualDirective,
+    AutofocusDirective
   ],
-  declarations: []
+  declarations: [
+    ValidateEqualDirective,
+    AutofocusDirective
+  ]
 })
 export class SharedModule { }