+++ src/app/test-management/create-test-suite-modal/create-test-suite-modal.component.ts
@@ -25,6 +25,7 @@ export class CreateTestSuiteModalComponent implements OnInit {
 
   createTestSuite() {
     this.testsuiteChange.emit(this.testsuite);
+    this.testsuite = new TestSuite();
   }
 
   ngOnInit() {